Install jupyter-{rsession,shiny}-proxy.
These are installed by repo2docker, but they get installed with the default python's (3.10) libraries. When we upgrade python in environment.yml, and when jupyter server launches from there, it doesn't see the previously installed proxies. So we have to install them ourselves.

Also update other libraries. This could have been done in a separate commit.